Move watchers to root of the src directory
