+++ packages/arbor-react/src/watchChildrenProps.test.ts
@@ -1,8 +1,8 @@
 import Arbor from "@arborjs/store"
 import { act, renderHook } from "@testing-library/react-hooks/native"
 
-import useArbor from "../useArbor"
-import { watchNodeProps } from "./watchNodeProps"
+import useArbor from "./useArbor"
+import { watchChildrenProps } from "./watchChildrenProps"
 
 interface Post {
   content: string
@@ -18,8 +18,8 @@ interface State {
   users: User[]
 }
 
-describe("watchNodeProps", () => {
-  it("does not update if mutation does not targets the given node", () => {
+describe("watchChildrenProps", () => {
+  it("does not update if mutation does not target any of the listed children props", () => {
     const store = new Arbor<State>({
       users: [
         { name: "Alice", age: 20, posts: [{ content: "Hello World" }] },
@@ -27,23 +27,42 @@ describe("watchNodeProps", () => {
       ]
     })
 
-    const { result } = renderHook(() => useArbor(store.root.users[0], watchNodeProps<User>("name")))
+    const { result } = renderHook(() => useArbor(store.root.users, watchChildrenProps<User>("name", "age")))
+
+    expect(result.all.length).toBe(1)
+
+    act(() => {
+      store.root.users[0].posts[0].content = "Updated post"
+      store.root.users[1].posts[0].content = "Another updated post"
+      store.root.users[0].posts = []
+      store.root.users[1].posts = []
+      store.root.users = []
+      store.setRoot({} as State)
+    })
+
+    expect(result.all.length).toBe(1)
+  })
+
+  it("updates when mutation targets the node being watched", () => {
+    const store = new Arbor<State>({
+      users: [
+        { name: "Alice", age: 20, posts: [{ content: "Hello World" }] },
+        { name: "Bob", age: 30, posts: [{ content: "My first post" }] },
+      ]
+    })
+
+    const { result } = renderHook(() => useArbor(store.root.users, watchChildrenProps<User>("name", "age")))
 
     expect(result.all.length).toBe(1)
 
     act(() => {
       store.root.users.push({ name: "Carol", age: 25, posts: [] })
-      store.root.users[0].posts.push({ content: "New post" })
-      store.root.users[0].posts[0].content = "Updated content"
-      store.root.users[1].name = "Bob Updated"
-      store.root.users[1].posts[0].content = "My first updated post"
-      store.root.users = []
     })
 
-    expect(result.all.length).toBe(1)
+    expect(result.all.length).toBe(2)
   })
 
-  it("does not update if mutation does not change any of the node props being watched", () => {
+  it("updates if mutation targets any of the listed children props", () => {
     const store = new Arbor<State>({
       users: [
         { name: "Alice", age: 20, posts: [{ content: "Hello World" }] },
@@ -51,31 +70,12 @@ describe("watchNodeProps", () => {
       ]
     })
 
-    const { result } = renderHook(() => useArbor(store.root.users[0], watchNodeProps<User>("name", "age")))
+    const { result } = renderHook(() => useArbor(store.root.users, watchChildrenProps<User>("name", "age")))
 
     expect(result.all.length).toBe(1)
 
     act(() => {
-      store.root.users[0].posts = []
-    })
-
-    expect(result.all.length).toBe(1)
-  })
-
-  it("updates when mutation targets the given node and any of the watched props change", () => {
-    const store = new Arbor<State>({
-      users: [
-        { name: "Alice", age: 20, posts: [{ content: "Hello World" }] },
-        { name: "Bob", age: 30, posts: [{ content: "My first post" }] },
-      ]
-    })
-
-    const { result } = renderHook(() => useArbor(store.root.users[0], watchNodeProps<User>("name", "age")))
-
-    expect(result.all.length).toBe(1)
-
-    act(() => {
-      store.root.users[0].name = "Alice Updated"
+      store.root.users[0].name = "Alice updated"
     })
 
     expect(result.all.length).toBe(2)
@@ -85,5 +85,17 @@ describe("watchNodeProps", () => {
     })
 
     expect(result.all.length).toBe(3)
+
+    act(() => {
+      store.root.users[1].name = "Bob updated"
+    })
+
+    expect(result.all.length).toBe(4)
+
+    act(() => {
+      store.root.users[1].age++
+    })
+
+    expect(result.all.length).toBe(5)
   })
 })
