+++ packages/arbor-react/src/watchChildrenProps.ts
@@ -7,6 +7,7 @@ export function watchChildrenProps<T extends object>(...props: PropOf<T>[]) {
   return <K extends object>(node: ArborNode<K>) =>
     (event: MutationEvent<K>) => {
       if (!isNode(node)) return false
+      if (event.mutationPath.is(node.$path)) return true
 
       const paths = props.map(prop => node.$path.child(":any").child(`#${prop}`).toString())
