+++ packages/arbor-react/src/watchNode.ts
