+++ packages/arbor-react/src/watchNodeProps.ts
