+++ packages/arbor-react/src/watchPaths.ts
